fixed dependecy issue and modified the view and added binding adapters

+++ app/src/main/java/com/anthonyra95/android/myworkouttracker/database/Exercise.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName ="exercise_table")
 data class Exercise (
     @PrimaryKey(autoGenerate = true)
-    var exerciseId: Long = 0L,
+    var exerciseId: Int = 0,
 
     @ColumnInfo(name = "exercise_name")
     var exerciseName: String = "",
