+++ app/src/main/java/com/anthonyra95/android/myworkouttracker/database/Workout.kt
@@ -10,7 +10,7 @@ data class Workout(
     var entryId: Long = 0L,
 
     @ColumnInfo(name = "exercise_id") //TODO make it a foreign key
-    var exerciseId: Long = 0L,
+    var exerciseId: Int = 0,
 
     @ColumnInfo(name = "start_time_milli")
     var startTimeMilli: Long = System.currentTimeMillis(),
